+++ skoolkit/config.py
@@ -101,13 +101,14 @@ COMMANDS = {
         'Time': (0, 'show_timings')
     },
     'skool2asm': {
+        'Address': ('', ''),
         'Base': (0, 'base'),
         'Case': (0, 'case'),
         'CreateLabels': (0, 'create_labels'),
         'EntryLabel': ('L{address}', ''),
         'EntryPointLabel': ('{main}_{index}', ''),
         'Quiet': (0, 'quiet'),
-        'Templates': ('', 'templates'),
+        'Templates': ('', ''),
         'Warnings': (1, 'warn')
     }
 }
@@ -155,10 +156,6 @@ def update_options(name, options, specs, config=None):
                     config[param] = value
             except ValueError:
                 pass
-    if config:
-        for param, (def_value, attr_name) in def_config.items():
-            if attr_name and not hasattr(options, attr_name):
-                setattr(options, attr_name, config[param])
 
 def show_config(section_name, config):
     print('[{}]'.format(section_name))